AnsonSSTVProtocol - Update commit - Hour 3
Attempting to decode the audio, but I can't figure out how to get the code to sync with the audio, hence it's a failed project (but experienceable !). Changes made to the encoding algorithm: Pixel separator changed to 700hz and 2200hz. Line separator and header code changed to 400hz and 700hz. Payload changed to 900hz and 1900hz.